+++ ddtrace/appsec/iast/taint_sinks/_base.py
@@ -9,9 +9,9 @@ from ddtrace.appsec._constants import IAST
 from ddtrace.appsec.iast import oce
 from ddtrace.appsec.iast._metrics import _set_metric_iast_executed_sink
 from ddtrace.appsec.iast._overhead_control_engine import Operation
-from ddtrace.appsec.iast._util import _has_to_scrub
-from ddtrace.appsec.iast._util import _is_evidence_value_parts
-from ddtrace.appsec.iast._util import _scrub
+from ddtrace.appsec.iast._utils import _has_to_scrub
+from ddtrace.appsec.iast._utils import _is_evidence_value_parts
+from ddtrace.appsec.iast._utils import _scrub
 from ddtrace.appsec.iast.reporter import Evidence
 from ddtrace.appsec.iast.reporter import IastSpanReporter
 from ddtrace.appsec.iast.reporter import Location
