+++ ddtrace/contrib/dbapi/__init__.py
@@ -4,7 +4,7 @@ Generic dbapi tracing code.
 import six
 
 from ddtrace import config
-from ddtrace.appsec.iast._util import _is_iast_enabled
+from ddtrace.appsec.iast._utils import _is_iast_enabled
 from ddtrace.internal.constants import COMPONENT
 
 from ...constants import ANALYTICS_SAMPLE_RATE_KEY
