+++ tests/appsec/iast/aspects/test_encode_decode_aspect.py
@@ -8,7 +8,7 @@ try:
     from ddtrace.appsec.iast._taint_tracking import OriginType
     from ddtrace.appsec.iast._taint_tracking import contexts_reset
     from ddtrace.appsec.iast._taint_tracking import create_context
-    from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
+    from ddtrace.appsec.iast._utils import _is_python_version_supported as python_supported_by_iast
 except (ImportError, AttributeError):
     pytest.skip("IAST not supported for this Python version", allow_module_level=True)
