+++ tests/appsec/iast/test_path_traversal.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
+from ddtrace.appsec.iast._utils import _is_python_version_supported as python_supported_by_iast
 from tests.appsec.iast.aspects.conftest import _iast_patched_module
