+++ tests/appsec/iast/test_taint_tracking.py
@@ -11,7 +11,7 @@ try:
     from ddtrace.appsec.iast._taint_tracking import taint_pyobject
     from ddtrace.appsec.iast._taint_tracking import taint_ranges_as_evidence_info
     from ddtrace.appsec.iast._taint_tracking.aspects import add_aspect
-    from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
+    from ddtrace.appsec.iast._utils import _is_python_version_supported as python_supported_by_iast
 except (ImportError, AttributeError):
     pytest.skip("IAST not supported for this Python version", allow_module_level=True)
