+++ tests/appsec/iast/test_weak_hash.py
@@ -3,7 +3,7 @@ import sys
 import pytest
 
 from ddtrace.appsec._constants import IAST
-from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
+from ddtrace.appsec.iast._utils import _is_python_version_supported as python_supported_by_iast
 from ddtrace.appsec.iast.constants import VULN_INSECURE_HASHING_TYPE
 from ddtrace.appsec.iast.taint_sinks.weak_hash import unpatch_iast
 from ddtrace.internal import core
