+++ tests/contrib/dbapi/test_dbapi_appsec.py
@@ -3,7 +3,7 @@ import pytest
 
 from ddtrace import Pin
 from ddtrace.appsec.iast import oce
-from ddtrace.appsec.iast._util import _is_python_version_supported
+from ddtrace.appsec.iast._utils import _is_python_version_supported
 from ddtrace.contrib.dbapi import TracedCursor
 from ddtrace.settings import Config
 from ddtrace.settings.integration import IntegrationConfig
