+++ tests/contrib/django/test_django_appsec_iast.py
@@ -8,7 +8,7 @@ from ddtrace import config
 from ddtrace.appsec._constants import IAST
 from ddtrace.appsec.iast import oce
 from ddtrace.appsec.iast._patch_modules import patch_iast
-from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
+from ddtrace.appsec.iast._utils import _is_python_version_supported as python_supported_by_iast
 from ddtrace.internal.compat import urlencode
 from tests.appsec.iast.iast_utils import get_line_and_hash
 from tests.utils import override_global_config
